+++ src/arrayOf.js
@@ -1,6 +1,0 @@
-import hasShape from './hasShape';
-import every from './every';
-
-const arrayOf = (shape) => every(hasShape(shape));
-
-export default arrayOf;